gpu-worker/requirements-base.txt
# Base requirements for all GPU workers
boto3>=1.34.0
redis>=5.0.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
